refactor(tasks): split task-detail-dialog into mode/feature modules
The 1,033-line TaskDetailDrawer (one component, ~26 interlocking hooks
across create/edit/detail modes) becomes a 54-line shell over six
cohesive modules, mirroring the models-page use-models-data precedent:

- use-task-detail.ts — the interlocking data layer: every hook, form
  draft, workflow/gate state, and action handler, returned as one typed
  object. The dense state web made a shared hook the sanctioned shape;
  useAgent stays LAST in the sequence so hook order is stable across
  every render path.
- task-detail-modes.tsx — the create/edit form + read-only detail view
  drawers plus the shared hero card (both consume the hook object).
- task-workflow-panels.tsx — workflow progress, approval-gate panel, and
  the edit-mode workflow preview (shared by both modes).
- task-notes-section.tsx — the notes block + dispatch-failure log panel
  leaf viewers (the two identical inline copies collapse into one).
- step-output-viewer.tsx — the prior-step output leaf viewers.

Behavior-preserving: JSX is byte-identical where it moved and the single
hook sequence lives in useTaskDetail, so no render path changes hook
order. The exported TaskDetailDrawer name + './task-detail-dialog'
specifier the kanban board imports are unchanged.

Fetch migration (WS3/P2 #18 straggler folded in): the workflow-
definitions load — a clean single-GET on mount — moves to useJsonFetch
(workflows derives via useMemo). The other three fetch effects stay raw:
the form-populate effect fires multiple conditional fetches plus field
resets; the workflowId-change definition fetch and the gate handlers all
write the same shared wfDefinition/wfInstance state, which useJsonFetch's
self-owned data state can't back; and fetchPriorOutput is an imperative
retry-poll loop. The local WorkflowInstance/WorkflowDefinition SDK-type
extensions move into use-task-detail.ts unchanged.

Test: the dialog test mocks @makinbakin/sdk/hooks, so it gains a
useJsonFetch stub with the standard { data, loading, error, refresh }
shape (no data — this test drives no workflow).
